# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SAVED4 OpenAPI Spec
  description: API specification for SAVED4
  version: 0.1.0
  x-fiddle-import-file: true
paths:
...
  /orders:
    get:
      operationId: getAllOrders
      summary: Get all orders
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
                title: GetAllOrdersOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - name: container_id
          in: query
          description: Filter orders by container identifier
          schema:
            type: integer
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: requester
          in: query
          description: Filter orders by the requester’s username
          schema:
            type: string
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: start_date
          in: query
          description: Filter orders created on or after this date (ISO 8601 format
            `YYYY-MM-DD`)
          schema:
            type: string
            format: date-time
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: end_date
          in: query
          description: Filter orders created on or before this date (ISO 8601 format
            `YYYY-MM-DD` )
          schema:
            type: string
            format: date-time
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: source
          in: query
          description: Filter orders by source location, identified by username
          schema:
            type: string
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: destination
          in: query
          description: Filter orders by destination location, identified by username
          schema:
            type: string
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: route_id
          in: query
          description: Filter orders by route identifier
          schema:
            type: integer
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: status
          in: query
          description: Filter orders by status (e.g., created, scheduled, undergoing,
            completed, canceled)
          schema:
            type: string
          required: false
          style: form
          explode: true
          allowReserved: false
        - name: operationType
          in: query
          description: Filter orders by operation type (e.g. loading, unloading)
          schema:
            type: string
          required: false
          style: form
          explode: true
          allowReserved: false
      description: Retrieves a list of orders from the system. Results are unordered.
      tags:
        - orders
      security: []
    post:
      summary: Create order
      operationId: createOrder
      description: Creates a new record of type Order.
      tags:
        - orders
      security: []
      responses:
        "201":
          description: The request was successful, and a new resource was created.
          content:
            application/json:
              schema:
                type: integer
                description: id of the newly created order
                title: CreateOrderCreated
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    delete:
      operationId: deleteOrder
      summary: Delete order
      parameters:
        - $ref: "#/components/parameters/order_id"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteOrderNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Deletes the company identified by the specified username.
      tags:
        - orders
      security: []
...
  /orders/{order_id}:
    get:
      operationId: getOrder
      tags:
        - orders
      summary: Get order
      parameters:
        - $ref: "#/components/parameters/order_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
                required:
                  - data
                title: GetOrderOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Get the order identified by order id
    patch:
      operationId: patchOrder
      summary: Update an existing order
      parameters:
        - $ref: "#/components/parameters/order_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
                required:
                  - data
                title: PutOrderOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Updates the details of an existing order identified by order id.
      tags:
        - orders
      security: []
      requestBody:
        required: true
        description: A JSON object containing the fields to be updated.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPartialUpdate"
...
components:
  parameters:
...
    order_id:
      name: order_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                  required:
                    - message
            required:
              - message
    Unauthorized:
      description: Authentication is required to access the requested resource. The
        client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
        Ensure the client has appropriate permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: The server cannot find the requested resource. The endpoint may be
        invalid or the resource may no longer exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
    Conflict:
      description: The request could not be completed due to a conflict with the
        current state of the resource. Resolve the conflict and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
  schemas:
...
    Order:
      type: object
      properties:
        order_id:
          type: integer
          description: Unique identifier of the order.
          example: 123
        requester:
          type: string
          description: User who requested the order.
          example: companyA
        made_at:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        source:
          type: string
          description: Username associated with the starting location of the order.
          example: companyA
        destination:
          type: string
          description: Username associated with the destination of the order.
          example: companyA
        route_id:
          type: integer
          description: ID of the route associated with the order.
          example: 12
        tw_start:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        tw_end:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        status:
          type: string
          enum:
            - created
            - scheduled
            - undergoing
            - completed
            - canceled
          example: created
        weight:
          type: number
          example: 12
        operation_type:
          type: string
          enum:
            - loading
            - unloading
          example: loading
        freight_value:
          type: number
          description: value of the contents in EUR.
          example: 12
        is_shared:
          type: boolean
        container_id:
          type: integer
        eta:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
      required:
        - order_id
        - requester
        - made_at
        - source
        - destination
        - route_id
        - tw_start
        - tw_end
        - status
        - weight
        - operation_type
        - freight_value
        - is_shared
        - container_id
        - eta
      description: An order made by a company
    OrderPartialUpdate:
      type: object
      properties:
        made_at:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        source:
          type: string
          description: Username associated with the starting location of the order.
          example: companyA
        destination:
          type: string
          description: Username associated with the destination of the order.
          example: companyA
        route_id:
          type: integer
          example: 132
        tw_start:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        tw_end:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        status:
          type: string
          enum:
            - created
            - scheduled
            - undergoing
            - completed
            - canceled
          example: created
        weight:
          type: number
          example: 12
        operation_type:
          type: string
          enum:
            - loading
            - unloading
          example: loading
        freight_value:
          type: number
          example: 1500
        container_id:
          type: integer
          example: 1
        eta:
          type: string
          format: date-time
        is_shared:
          type: boolean
      required: []
      description: Data transfer object for partially updating an existing Order
        (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: Order
...
    OrderSummary:
      type: object
      properties:
        order_id:
          type: integer
        made_at:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        source:
          type: string
          example: companyA
        destination:
          type: string
          example: companyA
        tw_start:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        tw_end:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        status:
          type: string
          enum:
            - created
            - scheduled
            - undergoing
            - completed
            - canceled
          example: created
        operation_type:
          type: string
          enum:
            - loading
            - unloading
          example: loading
        freight_value:
          type: number
          example: 342
        eta:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
      required:
        - order_id
        - made_at
        - source
        - destination
        - tw_start
        - tw_end
        - status
        - operation_type
        - freight_value
        - eta
      description: A simplified representation of a Order, typically used in list views.
      x-fiddle-dto-info:
        baseSchemaName: Order
...